+++ bootique-cayenne-test/src/test/java/io/bootique/cayenne/test/CayenneTestModuleIT.java
@@ -1,0 +1,34 @@
+package io.bootique.cayenne.test;
+
+import io.bootique.test.BQTestRuntime;
+import io.bootique.test.junit.BQTestFactory;
+import org.apache.cayenne.map.DataMap;
+import org.junit.Rule;
+import org.junit.Test;
+
+import static org.mockito.Matchers.any;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.verify;
+
+public class CayenneTestModuleIT {
+
+    @Rule
+    public BQTestFactory testFactory = new BQTestFactory();
+
+    @Test
+    public void testSchemaListeners() {
+
+        SchemaListener listener = mock(SchemaListener.class);
+
+        BQTestRuntime testRuntime = testFactory.app("-c", "classpath:config2.yml")
+                .autoLoadModules()
+                .module(b -> {
+                    CayenneTestModule.contributeSchemaListener(b).addBinding().toInstance(listener);
+                })
+                .createRuntime();
+
+
+        verify(listener).afterSchemaCreated(any(DataMap.class));
+    }
+
+}